Move vesync base entity to separate module
